refactor(project): fold the project scope default into the models package
The default lived in its own pkg/projectscope for two pure functions.
The separation bought nothing: pkg/models already holds the Project
type and DefaultProjectSlug, already exposes behaviour of exactly this
kind (GetOrganisationId, GetProjectObjectId), and is itself I/O-free,
so isolating a directory added no guarantee the doc comment did not
already carry. Both importers took pkg/models anyway, so the leaf
package reduced no dependency — it only added an import line.

The module boundary that does matter is kept: these stay in models
rather than moving next to the device ownership resolution in the
database module. Nine modules depend on models without database
(cli, ingest, vault, markers, queue, hub-anpr, hub-objecttracking,
hub-loitering, hub-pipeline-sequence), and making them take the mongo
driver to compute a value that must never be queried would invert the
point. Organisation ownership is verified against records and belongs
behind the driver; project ownership is defined and must not be.

Callers move from projectscope.X to models.X.
